refactor: Game/Player Type ( OptRcRef )

+++ src/utils/json.rs
@@ -3,13 +3,13 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct Item {
     pub id: String,
-    pub dbfid: i32,
+    pub dbfid: usize,
 }
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Card {
     pub id: String,
-    pub num: i32,
+    pub num: usize,
 }
 
 #[derive(Debug, Deserialize, Serialize)]
@@ -45,7 +45,7 @@ pub struct DeckCodes {
 #[derive(Debug, Deserialize)]
 pub struct GameConfigJson {
     pub DeckCodes: Vec<DeckCodes>,
-    pub Attacker: i32,
+    pub Attacker: usize,
     pub Names: Vec<Names>,
 }
 
@@ -63,12 +63,12 @@ pub struct GameConfigJson {
 #[derive(Debug, Deserialize, Serialize, Clone, Eq, PartialEq, Hash)]
 pub struct CardJson {
     pub id: Option<String>,
-    pub dbfid: Option<i32>,
-    pub cost: Option<i32>,
+    pub dbfid: Option<usize>,
+    pub cost: Option<usize>,
     pub name: Option<String>,
     pub text: Option<String>,
-    pub attack: Option<i32>,
-    pub health: Option<i32>,
+    pub attack: Option<usize>,
+    pub health: Option<usize>,
     pub collectible: Option<bool>,
 }
